bugs

+++ snd-g1.h
@@ -119,7 +119,7 @@ GtkWidget *get_help_menu_widget(void);
 
 void set_fft_window_beta(Float val);
 void set_fft_window_alpha(Float val);
-void set_transform_size(int val);
+void set_transform_size(off_t val);
 void set_fft_window(mus_fft_window_t val);
 void set_wavelet_type(int val);
 GtkWidget *fire_up_transform_dialog(bool managed);
